requirements.txt
Django==6.0.3
djangorestframework==3.14.0
drf-spectacular==0.27.2
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.8.2
dj-database-url==2.3.0
redis==5.0.1
django-redis==5.4.0
Pillow==10.2.0
celery==5.4.0
pytest==8.3.2
pytest-django==4.9.0
pytest-cov==5.0.0
